wip organize compiler config / defaults
