+++ packages/xmcp/src/compiler/get-webpack-config/plugins.ts
@@ -40,6 +40,8 @@ export class CreateTypeDefinitionPlugin {
         const xmcpConfig = getXmcpConfig();
 
         // Manually type the .xmcp/adapter/index.js file using a .xmcp/adapter/index.d.ts file
+
+        // TO DO add withAuth to the type definition & AuthConfig
         if (xmcpConfig.experimental?.adapter) {
           let typeDefinitionContent = "";
           if (xmcpConfig.experimental?.adapter == "nextjs") {
