+++ packages/xmcp/src/compiler/parse-xmcp-config.ts
@@ -1,85 +1,10 @@
 import fs from "fs";
 import path from "path";
-import { z } from "zod";
 import { webpack, type Configuration } from "webpack";
 import { createFsFromVolume, Volume } from "memfs";
 import { compilerContext } from "./compiler-context";
-
-export const DEFAULT_HTTP_PORT = 3002;
-export const DEFAULT_HTTP_BODY_SIZE_LIMIT = 1024 * 1024 * 10; // 10MB
-export const DEFAULT_HTTP_ENDPOINT = "/mcp";
-export const DEFAULT_HTTP_STATELESS = true;
-
-export const DEFAULT_TOOLS_DIR = "src/tools";
-
-// cors config schema
-const corsConfigSchema = z.object({
-  origin: z.union([z.string(), z.array(z.string()), z.boolean()]).optional(),
-  methods: z.union([z.string(), z.array(z.string())]).optional(),
-  allowedHeaders: z.union([z.string(), z.array(z.string())]).optional(),
-  exposedHeaders: z.union([z.string(), z.array(z.string())]).optional(),
-  credentials: z.boolean().optional(),
-  maxAge: z.number().optional(),
-});
-
-// oauth endpoints schema
-const oauthEndpointsSchema = z.object({
-  authorizationUrl: z.string(),
-  tokenUrl: z.string(),
-  revocationUrl: z.string().optional(),
-  userInfoUrl: z.string().optional(),
-  registerUrl: z.string(),
-});
-
-// oauth config schema
-const oauthConfigSchema = z.object({
-  endpoints: oauthEndpointsSchema,
-  issuerUrl: z.string(),
-  baseUrl: z.string(),
-  serviceDocumentationUrl: z.string().optional(),
-  pathPrefix: z.string().default("/oauth2"),
-  defaultScopes: z.array(z.string()).default(["openid", "profile", "email"]),
-});
-
-// adapter config schema
-const adapterConfigSchema = z.enum(["express", "nextjs"]);
-
-// experimental features schema
-const experimentalConfigSchema = z.object({
-  oauth: oauthConfigSchema.optional(),
-  adapter: adapterConfigSchema.optional(),
-});
-
-// paths config
-const pathsConfigSchema = z.object({
-  tools: z.string().default(DEFAULT_TOOLS_DIR),
-  // to do add resources prompts etc
-});
-
-// TODO extract all this config and schemas to a separate file
-const configSchema = z.object({
-  stdio: z.boolean().optional(),
-  http: z
-    .union([
-      z.boolean(),
-      z.object({
-        port: z.number().default(DEFAULT_HTTP_PORT),
-        bodySizeLimit: z.number().default(DEFAULT_HTTP_BODY_SIZE_LIMIT),
-        debug: z.boolean().default(false),
-        endpoint: z.string().default(DEFAULT_HTTP_ENDPOINT),
-        cors: corsConfigSchema.optional(),
-      }),
-    ])
-    .optional(),
-  experimental: experimentalConfigSchema.optional(),
-  paths: pathsConfigSchema.optional().default({
-    tools: DEFAULT_TOOLS_DIR,
-  }),
-  webpack: z.function().args(z.any()).returns(z.any()).optional(),
-});
-
-type InputSchema = z.input<typeof configSchema>;
-type OutputSchema = z.output<typeof configSchema>;
+import { configSchema, type InputSchema, type OutputSchema } from "./config";
+import { DEFAULT_TOOLS_DIR } from "./config/constants";
 
 /** Config type for the user to provide */
 export type XmcpInputConfig = Omit<InputSchema, "webpack"> & {
